[identity-split] split disentangle-identity into identity + economy crates
Mechanical refactor isolating experimental economic primitives
(commons_pool, intent, agreement, oracle) into a new
disentangle-economy crate, keeping the protocol-tier identity
surface minimal (did, capability, coherence, petname, governance,
graph, transactions, proposal).

- Create disentangle-economy crate and register in workspace.
- Move the four modules with git history preserved via git mv.
- Drop the moved modules from disentangle-identity/src/lib.rs.
- Split the identity integration test: oracle_* tests move to
  disentangle-economy/tests/oracle_integration.rs; governance,
  capability, and coherence tests stay in identity.
- Redirect disentangle-node consumers to the new paths.
