+++ disentangle-economy/src/agreement.rs
