+++ disentangle-economy/src/intent.rs
